requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
asyncpg==0.30.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
flatbuffers==25.9.23
frozenlist==1.8.0
greenlet==3.2.4
humanfriendly==10.0
idna==3.11
imageio==2.37.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
netaddr==1.3.0
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.0
psycopg==3.2.11
psycopg-binary==3.2.11
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyMatting==1.1.14
python-dotenv==1.1.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
SQLAlchemy==2.0.44
yookassa==3.7.1
